ecom_agent/prototype/requirements.txt
openai==1.75.0
gradio==5.25.2
python-dotenv==1.1.0
pillow==10.2.0
requests==2.31.0
